requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.0
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
numpy==2.3.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
